requirements.txt
beautifulsoup4==4.12.0
soupsieve==2.4
python-dotenv==1.0.0
